auth/requirements.txt
alembic==1.12.0
async-fastapi-jwt-auth==0.6.1
async-sqlalchemy==1.0.0
databases==0.8.0
email-validator==2.0.0
fastapi==0.103.0
pydantic==2.3.0
pydantic-settings==2.0.3
uvicorn==0.23.2
